backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic[email]==2.6.3
sqlalchemy==2.0.28
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
cors-headers==0.1.0
alembic==1.13.1
pytest==8.0.2
httpx==0.27.0
